requirements_local.txt
Django==1.10.1
dj-database-url==0.4.1
django-celery==3.2.1
-e git+https://github.com/sigmavirus24/github3.py.git@bbed706433db4c37fbc826f8eab92701b0be5bea#egg=github3.py
-e git+https://github.com/elsys/elsys-teachers-tools.git@686762174eef487c21246a679a4a068c39d8d679#egg=elsys_teachers_tools
GitPython==2.0.8
gspread==0.4.1
gunicorn==19.6.0
whitenoise==3.2.2
redis==2.10.5
django-cors-headers==1.2.0
oauth2client==3.0.0
pyOpenSSL==16.1.0
PyYAML==3.12
